Rename modules in plural form
