--- src/api/__init__.py
@@ -1,2 +1,0 @@
-from api.datadragon import DataDragon
-from api.lolfandom import LolFandom
