+++ src/apis/__init__.py
@@ -1,0 +1,2 @@
+from apis.datadragon import DataDragon
+from apis.lolfandom import LolFandom
