--- src/constant/__init__.py
@@ -1,4 +1,0 @@
-from constant.embedded import Embedded
-from constant.monsoon import Monsoon
-from constant.stylesheet import Stylesheet
-from constant.websocketevent import WebSocketEvent
