+++ src/constants/__init__.py
@@ -1,0 +1,4 @@
+from constants.embedded import Embedded
+from constants.monsoon import Monsoon
+from constants.stylesheet import Stylesheet
+from constants.websocketevent import WebSocketEvent
