+++ src/constants/embedded.py
