+++ src/constants/websocketevent.py
