--- src/controller/__init__.py
@@ -1,2 +1,0 @@
-from controller.eventdatacontroller import EventDataController
-from controller.leagueclientcontroller import LeagueClientController
