+++ src/controllers/__init__.py
@@ -1,0 +1,2 @@
+from controllers.eventdatacontroller import EventDataController
+from controllers.leagueclientcontroller import LeagueClientController
