--- src/model/__init__.py
@@ -1,2 +1,0 @@
-from model.balancemodel import BalanceModel
-from model.sessionmodel import SessionModel
