--- src/service/__init__.py
@@ -1,2 +1,0 @@
-from service.executorservice import ExecutorService
-from service.timerservice import TimerService
