--- src/template/__init__.py
@@ -1,1 +1,0 @@
-from template.systemtray import SystemTray
