--- src/view/__init__.py
@@ -1,2 +1,0 @@
-from view.mainview import MainView
-from view.aboutview import AboutView
