+++ src/views/__init__.py
@@ -1,0 +1,2 @@
+from views.mainview import MainView
+from views.aboutview import AboutView
